// web/src/api/models/TypesFileDiff.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TypesDiffLine } from './TypesDiffLine';
import {
    TypesDiffLineFromJSON,
    TypesDiffLineFromJSONTyped,
    TypesDiffLineToJSON,
    TypesDiffLineToJSONTyped,
} from './TypesDiffLine';

/**
 * 
 * @export
 * @interface TypesFileDiff
 */
export interface TypesFileDiff {
    /**
     * 
     * @type {string}
     * @memberof TypesFileDiff
     */
    file: string;
    /**
     * 
     * @type {Array<TypesDiffLine>}
     * @memberof TypesFileDiff
     */
    lines?: Array<TypesDiffLine>;
    /**
     * 
     * @type {string}
     * @memberof TypesFileDiff
     */
    status: TypesFileDiffStatusEnum;
}


/**
 * @export
 */
export const TypesFileDiffStatusEnum = {
    Added: 'added',
    Removed: 'removed',
    Modified: 'modified'
} as const;
export type TypesFileDiffStatusEnum = typeof TypesFileDiffStatusEnum[keyof typeof TypesFileDiffStatusEnum];


/**
 * Check if a given object implements the TypesFileDiff interface.
 */
export function instanceOfTypesFileDiff(value: object): value is TypesFileDiff {
    if (!('file' in value) || value['file'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function TypesFileDiffFromJSON(json: any): TypesFileDiff {
    return TypesFileDiffFromJSONTyped(json, false);
}

export function TypesFileDiffFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesFileDiff {
    if (json == null) {
        return json;
    }
    return {
        
        'file': json['file'],
        'lines': json['lines'] == null ? undefined : ((json['lines'] as Array<any>).map(TypesDiffLineFromJSON)),
        'status': json['status'],
    };
}

export function TypesFileDiffToJSON(json: any): TypesFileDiff {
    return TypesFileDiffToJSONTyped(json, false);
}

export function TypesFileDiffToJSONTyped(value?: TypesFileDiff | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'file': value['file'],
        'lines': value['lines'] == null ? undefined : ((value['lines'] as Array<any>).map(TypesDiffLineToJSON)),
        'status': value['status'],
    };
}